Function that prints a singly linked list

+++ 0x12-singly_linked_lists/lists.h
@@ -17,7 +17,7 @@
 typedef struct list_s
 {
     char *str;
-    size_t int len;
+    unsigned int len;
     struct list_s *next;
 } list_t;
